Time: 2 ms (69.82%) | Memory: 41.7 MB (58.03%) - LeetSync

+++ 389-find-the-difference/find-the-difference.java
@@ -2,10 +2,10 @@ class Solution {
     public char findTheDifference(String s, String t) {
         int sum=0;
         for(int i=0;i<t.length();i++){
-            sum = sum+t.charAt(i);
+            sum = sum+(int)t.charAt(i);
         }
         for(int i=0;i<s.length();i++){
-            sum = sum-s.charAt(i);
+            sum = sum-(int)s.charAt(i);
         }
         return (char)sum;
     }
